+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/environment-check
 
+## 0.16.5
+
+### Patch Changes
+
+-   Updated dependencies [b458bf43]
+    -   @sap-ux/project-access@1.15.0
+    -   @sap-ux/axios-extension@1.7.1
+
 ## 0.16.4
 
 ### Patch Changes